restructured main file and added term. arg. readin
