--- molara.py
@@ -1,43 +1,0 @@
-#!/usr/bin/env python3
-
-import sys
-
-from PySide6.QtGui import QSurfaceFormat
-from PySide6.QtWidgets import QApplication, QMainWindow, QFileDialog
-
-# Important:
-# You need to run the following command to generate the ui_form.py file
-#     pyside6-uic form.ui -o ui_form.py, or
-#     pyside2-uic form.ui -o ui_form.py
-from Gui.ui_form import Ui_MainWindow
-from Molecule.Molecule import read_xyz
-
-format = QSurfaceFormat()
-format.setVersion(4, 1)
-format.setProfile(QSurfaceFormat.CoreProfile)
-QSurfaceFormat.setDefaultFormat(format)
-
-
-class MainWindow(QMainWindow):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.ui = Ui_MainWindow()
-        self.ui.setupUi(self)
-
-    def show_xyz(self):
-        fileName = QFileDialog.getOpenFileName(self, "Open .xyz file", "/home", "Image Files (*.xyz)")
-        mol = read_xyz(fileName[0])
-        widget.ui.openGLWidget.set_molecule(mol)
-
-
-if __name__ == "__main__":
-    app = QApplication(sys.argv)
-    widget = MainWindow()
-    widget.setWindowTitle('Molara')
-    widget.show()
-    widget.ui.action_xyz.triggered.connect(widget.show_xyz)
-    widget.ui.actionReset_View.triggered.connect(widget.ui.openGLWidget.reset_view)
-    widget.ui.actionDraw_Axes.triggered.connect(widget.ui.openGLWidget.toggle_axes)
-    widget.ui.actionCenter_Molecule.triggered.connect(widget.ui.openGLWidget.center_molecule)
-    widget.ui.quit.triggered.connect(widget.close)
-    sys.exit(app.exec())
